+++ Code/Data Classes/AbilityData.cs
@@ -1,7 +1,6 @@
 ﻿
 namespace JLPlugin.Data
 {
-    [System.Serializable]
     public class AbilityData
     {
         public string name;