+++ Code/Data Classes/EncounterData.cs
@@ -1,12 +1,9 @@
-﻿using APIPlugin;
-using System;
+﻿
 using System.Collections.Generic;
-using static DiskCardGame.EncounterBlueprintData;
 
 namespace JLPlugin.Data
 {
-    [System.Serializable]
-    public partial class CustomEncounterData
+    public partial class EncounterData
     {
         public string name;