+++ Code/Data Classes/SpecialAbilityData.cs
@@ -1,7 +1,6 @@
 ﻿
 namespace JLPlugin.Data
 {
-    [System.Serializable]
     public class SpecialAbilityData
     {
         public string name;